Stone-Miner-Builder-Only-GPU
GPU Multi coin Mining
Windows defender is an exception.
WatchDog
Autorun
Bypassing the UAC.
The miner does not work when the task manager is open
Checking the operation of the miner (attempt to restart in case of inactivity).
Determination of the full screen - the miner will be temporarily suspended during the game
x64 bit depth